requirements.txt
# RepoReadme - Automatic README Generator
# Core dependencies for repository analysis and README generation

# GitHub/GitLab API access
PyGithub>=1.55.0
python-gitlab>=3.0.0
requests>=2.28.0

# GUI framework
# tkinter is included with Python by default

# File processing and parsing
PyYAML>=6.0
toml>=0.10.0
python-frontmatter>=1.0.0

# Markdown processing and generation
markdown>=3.4.0
jinja2>=3.0.0

# Code analysis and detection
gitpython>=3.1.0

# Template and documentation
mistune>=2.0.0
pygments>=2.10.0

# Utility libraries
python-dateutil>=2.8.0
colorama>=0.4.0

# Optional dependencies (install if needed)
# pandas>=2.0.0
# tabulate>=0.9.0
# pillow>=10.0.0
# beautifulsoup4>=4.12.0
# lxml>=4.9.0

# Development and testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0